Use UUID's to fetch journey data for edit
UUID’s should be used to handle journey posts because they’re
impossible (probably not entirely, but still) to guess like the ‘id’
sequence of numbers which strengthens security.

Also, the ‘data-journey-uuid’ attribute will only be written to the
HTML if the authenticated traveler is the owner of the particular post,
ensuring we don’t hand out the UUID’s to every journey post
unnecessarily, and meaning travelers should only have the UUID’s of
their own posts which prevents them from editing the posts of others
without permission.